Correct initial set up of MCMC analysis.
This has a better understanding and a better choice of values, for
the initial course "grid-like" search, the finner resulting MCMC
search (including more accurate staring positions and burnin length).
Finally there is a better understanding of emcee's step method and
the need for more walkers.

The result should only marginally increase the time it takes to run
a complete run, minimize the time to debug the code, but add some
significant time on a burnin analysis run.